+++ packages/cli/__tests__/zosfiles/__unit__/search/ds/Datasets.handler.unit.test.ts
@@ -98,7 +98,7 @@ describe("Search Datasets handler", () => {
 
             expect(Search.dataSets).toHaveBeenCalledTimes(1);
             expect(Search.dataSets).toHaveBeenCalledWith(fakeSession, {
-                dsn: pattern,
+                pattern,
                 searchString,
                 caseSensitive: undefined,
                 mainframeSearch: undefined,